image_loader 모듈에 있는 내용을 file_io 모듈애 포함
