+++ tests/file_io.py
@@ -8,8 +8,29 @@ from scipy import io
 import numpy as np
 import h5py
 import random
+import re
 
 random.seed(111)
+
+class FileSorter:
+    def __init__(self):
+        pass
+    
+    def sort(self, list_of_strs):
+        list_of_strs.sort(key=self._alphanum_key)
+
+    def _tryint(self, s):
+        try:
+            return int(s)
+        except:
+            return s
+    
+    def _alphanum_key(self, s):
+        """ Turn a string into a list of string and number chunks.
+            "z23a" -> ["z", 23, "a"]
+        """
+        return [ self._tryint(c) for c in re.split('([0-9]+)', s) ]
+
 
 class File(object):
     __metaclass__ = abc.ABCMeta
@@ -49,7 +70,7 @@ class FileJson(File):
         """
         return json.loads(open(filename).read())
     
-    # Todo : Exception Ã³¸®
+    # Todo : Exception Ã³ï¿½ï¿½
     def write(self, data, filename, write_mode="w"):
         self._check_directory(filename)        
         with open(filename, write_mode) as f:
@@ -112,7 +133,7 @@ class FileHDF5(File):
         db.close()
 
 
-def list_files(directory, pattern="*.*", n_files_to_sample=None, recursive_option=True):
+def list_files(directory, pattern="*.*", n_files_to_sample=None, recursive_option=True, random_order=True):
     """list files in a directory matched in defined pattern.
 
     Parameters
@@ -150,10 +171,15 @@ def list_files(directory, pattern="*.*", n_files_to_sample=None, recursive_optio
         for p in glob.glob(os.path.join(dir_, pattern)):
             files.append(p)
     
+    FileSorter().sort(files)
+        
     if n_files_to_sample is not None:
-        files = random.sample(files, n_files_to_sample)
-
+        if random_order:
+            files = random.sample(files, n_files_to_sample)
+        else:
+            files = files[:n_files_to_sample]
     return files
+
 
 if __name__ == "__main__":
     import doctest
